lines2insns: just skip debug info entries w/o section info
It seems possible for a debug info entry for a given source location
have no section info. Or, at least, lines2insns could not find it in
some of my experiments with test_common_target.ko on Fedora 25 x64,
GCC 6.3.1.

Namely, I tried to use it for cfake.c:348:

344 static void __exit
345 cfake_exit_module(void)
346 {
347 	cfake_cleanup_module(cfake_ndevices);
348 	return; // <<< this line
349 }

dwfl_lineinfo() actually found something for that source location,
Dwarf_Addr was 1464. The sections, however, had the following locations:
[DBG] section .text, addr: 48, size: 679
[DBG] section .text.unlikely, addr: 727, size: 163
[DBG] section .init.text, addr: 890, size: 557
[DBG] section .exit.text, addr: 1447, size: 17

That is, that Dwarf_Addr pointed right behind the section .exit.text.
Not sure, if this the intended behaviour of GCC/libdw/whatever.

Let us just skip such cases rather than throw an error and exit like
it was before.