core: Silence "BP has already been disabled" message by default
There are legit ways to trigger a bunch of such messages. For example,
if the kernel does not use full preemption and each CPU is waiting in a
delay in the SW BP handler (preemption is disabled there), the timer
that sets a HW BP may fire too late.

In this case, HW BP will not be set at all, it will just be marked as
disabled. The subsequent attempt to clear the HW BP will output the
message that the HW BP is already disabled.