lines2insns: If there are no data a source file, try without the path
Debug info may be associated with a full path to the source file or,
it seems, with the base name of that file. I cannot predict which of
the variants will be used on a given system, so let us check both. The
original path is checked first, then the base name, if it is different.