core: get rid of hwbp::usage_count altogether
Come to think of it, usage_count was only needed to see if some CPU was using
the HW BP or no CPUs were using it.

This can be easily determined from the attributes of the underlying Perf
events. Let us remove usage_count then and avoid confusion it causes.