core: fix specification of MOVSXD instruction
This 64-bit form of MOVSX with opcode 0x63 always reads 4 bytes (type "Ed")
rather than 8 ("Ev").

KASAN spotted that RaceHound tried to read 8 bytes of "cfake_ndevices"
variable in the tests (for repeated-read checks while it was only 4 bytes
in size. This was because of that incorrect information about MOVSXD in
the opcode table.

Fix that.