lines2isns: Handle debug info from separate files as well
It is quite common to keep debug info for a kernel module in a separate
file rather than in the same file where the code is. Consider,
distro-specific kernels and their debug info packages, for example.

lines2insns is now able to handle that, see lines2insns --help for
details.