Do not use in_atomic(): it is pointless
in_atomic() always returns 0 if CONFIG_PREEMPT_COUNT is not set, which
is almost always the case (Ubuntu with a generic kernel, VZ7, ...).

RaceHound cannot reliably detect if an instruction runs in atomic
context, e.g., under spinlock.

Let us do a different thing instead. Do not make delays in non-task
context (IRQ, SoftIRQ, NMI) at all. RaceHound is more likely to find
something with longer delays, which is not how IRQs should work, for
instance. Let us play safer.

RaceHound is still unable to detect spinlocks and disabled preemption in
many cases. This could be alleviated by using other tools to filter
the instructions to be monitored before feeding them to RaceHound. This
mostly affects sweeping, though. If RaceHound is used to check if a race
found by some other means (e.g. KEDR) actually happens, it is needed to
analyze the racing parts of the code beforehand anyway. So, that
scenario should be OK.